backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
supabase==2.4.2
pandas==2.2.1
pyarrow==15.0.2
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
reportlab==4.1.0
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-magic==0.4.27
groq==0.5.0
modal==0.62.231
numpy==1.26.4
